+++ packages/core/src/index.ts
@@ -4,3 +4,5 @@ export * from './lib/image-upload.service';
 export * from './lib/field-registry.service';
 export * from './lib/async-validator-registry.service';
 export * from './lib/form-persistence.service';
+export * from './lib/providers';
+export * from './lib/presets';